Add HFConverterCallback for end-of-training HuggingFace conversion
Refactor the HF conversion logic out of the example CLI script into
olmo_core.nn.hf.convert_checkpoint so it can be reused from a training
callback. The new HFConverterCallback gathers the full model state dict
at the end of training and converts it to HuggingFace format on rank 0.
