+++ src/olmo_core/nn/hf/__init__.py
@@ -18,8 +18,14 @@ from .convert import (
     get_converter_from_hf,
     get_converter_to_hf,
 )
+from .convert_checkpoint import (
+    convert_checkpoint_to_hf,
+    load_config,
+    validate_conversion,
+)
 
 __all__ = [
+    "convert_checkpoint_to_hf",
     "convert_hybrid_state_to_hf",
     "convert_state_from_hf",
     "convert_state_to_hf",
@@ -29,7 +35,9 @@ __all__ = [
     "get_hybrid_hf_config",
     "get_hybrid_layer_types",
     "is_olmo_hybrid_model",
+    "load_config",
     "load_hf_model",
     "save_hf_hybrid_model",
     "save_hf_model",
+    "validate_conversion",
 ]
